Remove Strategic IDAM qualifiers and made it default instead
