+++ infrastructure/aat.tfvars
@@ -1,6 +1,5 @@
 vault_env = "preprod"
 idam_api_baseurl = "https://idam-api.aat.platform.hmcts.net"
-idam_strategic_enabled = "true"
 
 capacity = "2"
