+++ infrastructure/prod.tfvars
@@ -1,6 +1,5 @@
 vault_env = "prod"
 idam_api_baseurl = "https://idam-api.platform.hmcts.net"
-idam_strategic_enabled = "true"
 
 capacity = "2"
 documentation_swagger_enabled = "false"
