+++ infrastructure/saat.tfvars
@@ -1,3 +1,2 @@
 vault_env = "test"
 idam_api_baseurl = "http://idam-api-idam-saat.service.core-compute-idam-saat.internal"
-idam_strategic_enabled = "true"
