+++ src/main/java/uk/gov/hmcts/reform/divorce/orchestration/tasks/AuthenticateRespondent.java
@@ -3,7 +3,6 @@ package uk.gov.hmcts.reform.divorce.orchestration.tasks;
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.reform.divorce.orchestration.client.IdamClient;
 import uk.gov.hmcts.reform.divorce.orchestration.domain.model.idam.UserDetails;
@@ -16,7 +15,6 @@ import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.Orchestrati
 @Component
 public class AuthenticateRespondent implements Task<Boolean> {
     @Autowired
-    @Qualifier("idamClient")
     private IdamClient idamClient;
 
     @Autowired
