+++ src/main/java/uk/gov/hmcts/reform/divorce/orchestration/util/AuthUtil.java
@@ -2,7 +2,6 @@ package uk.gov.hmcts.reform.divorce.orchestration.util;
 
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.reform.divorce.orchestration.client.IdamClient;
@@ -45,7 +44,7 @@ public class AuthUtil {
     private final IdamClient idamClient;
 
     @Autowired
-    public AuthUtil(@Qualifier("idamClient") IdamClient idamClient) {
+    public AuthUtil(IdamClient idamClient) {
         this.idamClient = idamClient;
     }
