+++ src/test/java/uk/gov/hmcts/reform/divorce/orchestration/functionaltest/IdamTestSupport.java
@@ -90,9 +90,6 @@ public abstract class IdamTestSupport extends MockedFunctionalTest {
     @Value("${idam.caseworker.password}")
     private String caseworkerPassword;
 
-    @Value("${idam.strategic.enabled}")
-    private boolean sidamEnabled;
-
     void stubUserDetailsEndpoint(HttpStatus status, String authHeader, String message) {
         idamServer.stubFor(get(IDAM_USER_DETAILS_CONTEXT_PATH)
             .withHeader(AUTHORIZATION, new EqualToPattern(authHeader))
@@ -153,28 +150,15 @@ public abstract class IdamTestSupport extends MockedFunctionalTest {
 
     void stubPinAuthoriseEndpoint(HttpStatus status, String responseBody)
         throws UnsupportedEncodingException {
-
-        if (sidamEnabled) {
-            idamServer.stubFor(get(IDAM_PIN_DETAILS_CONTEXT_PATH
-                    + "?client_id=" + authClientId
-                    + "&redirect_uri=" + URLEncoder.encode(authRedirectUrl, StandardCharsets.UTF_8.name()))
-                .withHeader("pin", new EqualToPattern(TEST_PIN))
-                .willReturn(aResponse()
-                    .withStatus(status.value())
-                    .withHeader(CONTENT_TYPE, APPLICATION_JSON_VALUE)
-                    .withHeader(LOCATION_HEADER, PIN_AUTH_URL_WITH_REDIRECT)
-                    .withBody(responseBody)));
-        } else {
-            idamServer.stubFor(post(IDAM_AUTHORIZE_CONTEXT_PATH
-                + "?response_type=" + CODE
-                + "&client_id=" + authClientId
+        idamServer.stubFor(get(IDAM_PIN_DETAILS_CONTEXT_PATH
+                + "?client_id=" + authClientId
                 + "&redirect_uri=" + URLEncoder.encode(authRedirectUrl, StandardCharsets.UTF_8.name()))
-                .withHeader(AUTHORIZATION, new EqualToPattern(PIN_AUTHORIZATION))
-                .willReturn(aResponse()
-                    .withStatus(status.value())
-                    .withHeader(CONTENT_TYPE, APPLICATION_JSON_VALUE)
-                    .withBody(responseBody)));
-        }
+            .withHeader("pin", new EqualToPattern(TEST_PIN))
+            .willReturn(aResponse()
+                .withStatus(status.value())
+                .withHeader(CONTENT_TYPE, APPLICATION_JSON_VALUE)
+                .withHeader(LOCATION_HEADER, PIN_AUTH_URL_WITH_REDIRECT)
+                .withBody(responseBody)));
     }
 
     void stubTokenExchangeEndpoint(HttpStatus status, String authCode, String responseBody)
