+++ src/test/java/uk/gov/hmcts/reform/divorce/orchestration/tasks/ValidateExistingSolicitorLinkTest.java
@@ -5,7 +5,7 @@ import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
-import uk.gov.hmcts.reform.divorce.orchestration.client.StrategicIdamClient;
+import uk.gov.hmcts.reform.divorce.orchestration.client.IdamClient;
 import uk.gov.hmcts.reform.divorce.orchestration.domain.model.ccd.CaseDetails;
 import uk.gov.hmcts.reform.divorce.orchestration.domain.model.idam.UserDetails;
 import uk.gov.hmcts.reform.divorce.orchestration.framework.workflow.task.DefaultTaskContext;
@@ -27,7 +27,7 @@ public class ValidateExistingSolicitorLinkTest {
     private static final String TOKEN = "token";
 
     @Mock
-    private StrategicIdamClient idamClient;
+    private IdamClient idamClient;
 
     @InjectMocks
     private ValidateExistingSolicitorLink validateCaseData;
