+++ src/test/resources/application.yml
@@ -303,8 +303,6 @@ idam:
   caseworker:
     username: dummy
     password: dummy
-  strategic:
-    enabled: true
   s2s-auth:
     url: ${SERVICE_AUTH_PROVIDER_URL:http://localhost:4504}
     microservice: ${SERVICE_AUTH_MICROSERVICE:divorce_frontend}
